refactor: test 용 객체들 패키지 이동
