+++ backend/src/main/java/com/stampcrush/backend/auth/test/ManagerAuthTestController.java
@@ -1,30 +1,30 @@
-package com.stampcrush.backend.auth.api;
+package com.stampcrush.backend.auth.test;
 
-import com.stampcrush.backend.auth.api.request.OAuthRegisterCustomerCreateRequest;
+import com.stampcrush.backend.auth.api.request.OAuthRegisterOwnerCreateRequest;
 import com.stampcrush.backend.auth.api.response.AuthTokensResponse;
-import com.stampcrush.backend.auth.application.VisitorAuthTestService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@Profile("test")
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/api/login")
-public class VisitorAuthTestController {
+@RequestMapping("/api/admin/login")
+public class ManagerAuthTestController {
 
-    private final VisitorAuthTestService visitorAuthTestService;
+    private final ManagerAuthTestService managerAuthTestService;
 
     @PostMapping("/register/test/token")
-    public ResponseEntity<AuthTokensResponse> joinRegisterCustomer(
-            @RequestBody OAuthRegisterCustomerCreateRequest request
+    public ResponseEntity<AuthTokensResponse> joinManager(
+            @RequestBody OAuthRegisterOwnerCreateRequest request
     ) {
         return ResponseEntity.ok().body(
-                visitorAuthTestService.join(
+                managerAuthTestService.join(
                         request.getNickname(),
-                        request.getEmail(),
                         request.getoAuthProvider(),
                         request.getoAuthId()
                 )
