+++ backend/src/main/java/com/stampcrush/backend/auth/test/VisitorAuthTestController.java
@@ -1,29 +1,31 @@
-package com.stampcrush.backend.auth.api;
+package com.stampcrush.backend.auth.test;
 
-import com.stampcrush.backend.auth.api.request.OAuthRegisterOwnerCreateRequest;
+import com.stampcrush.backend.auth.api.request.OAuthRegisterCustomerCreateRequest;
 import com.stampcrush.backend.auth.api.response.AuthTokensResponse;
-import com.stampcrush.backend.auth.application.ManagerAuthTestService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@Profile("test")
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/api/admin/login")
-public class ManagerAuthTestController {
+@RequestMapping("/api/login")
+public class VisitorAuthTestController {
 
-    private final ManagerAuthTestService managerAuthTestService;
+    private final VisitorAuthTestService visitorAuthTestService;
 
     @PostMapping("/register/test/token")
-    public ResponseEntity<AuthTokensResponse> joinManager(
-            @RequestBody OAuthRegisterOwnerCreateRequest request
+    public ResponseEntity<AuthTokensResponse> joinRegisterCustomer(
+            @RequestBody OAuthRegisterCustomerCreateRequest request
     ) {
         return ResponseEntity.ok().body(
-                managerAuthTestService.join(
+                visitorAuthTestService.join(
                         request.getNickname(),
+                        request.getEmail(),
                         request.getoAuthProvider(),
                         request.getoAuthId()
                 )
